Made changes to build settings
Assets are now more reasonable and are placed next to the executable in the folder structure. This is reflected in the main file.

The folder you would distribute now appears next to the README, LICENSE and gitignore file in the Packaged folder. Now compression as of now, but this may be changed in the future to automatically compress into zip files.