Work on the input system
I have to admit, this would be the point where setting up branches would be appropriate. I just don't feel like doing that because no one is using this engine, so the main branch doesn't have to work.

So yeah, as of this comment the input system is still very much in progress and not working.