Structuring changes to make usage easier
IceCoreV2 was difficult to work with and contribute to at the same time as working on my own project. So I made some minor changes to file structure to allow IceCore to be a submodule in my git repo for my game.

This is how I would recommend using IceCore in any other projects as well